# 🏗🚀 Speed up pre-closure `babel` transforms