--- a/build-system/compile/compile.js
+++ b/build-system/compile/compile.js
@@ -31,9 +31,10 @@ const {
 } = require('./closure-compile');
 const {checkForUnknownDeps} = require('./check-for-unknown-deps');
 const {checkTypesNailgunPort, distNailgunPort} = require('../tasks/nailgun');
-const {CLOSURE_SRC_GLOBS, SRC_TEMP_DIR} = require('./sources');
+const {CLOSURE_SRC_GLOBS} = require('./sources');
 const {isTravisBuild} = require('../common/travis');
 const {postClosureBabel} = require('./post-closure-babel');
+const {preClosureBabel} = require('./pre-closure-babel');
 const {singlePassCompile} = require('./single-pass');
 const {VERSION: internalRuntimeVersion} = require('./internal-version');
 
@@ -45,17 +46,6 @@ let inProgress = 0;
 // during various local development scenarios.
 // See https://github.com/google/closure-compiler-npm/issues/9
 const MAX_PARALLEL_CLOSURE_INVOCATIONS = isTravisBuild() ? 4 : 1;
-
-/**
- * Prefixes the tmp directory if we need to shadow files that have been
- * preprocessed by babel in the `dist` task.
- *
- * @param {!Array<string>} paths
- * @return {!Array<string>}
- */
-function convertPathsToTmpRoot(paths) {
-  return paths.map(path => path.replace(/^(!?)(.*)$/, `$1${SRC_TEMP_DIR}/$2`));
-}
 
 // Compiles AMP with the closure compiler. This is intended only for
 // production use. During development we intend to continue using
@@ -358,8 +348,6 @@ function compile(
       delete compilerOptions.define;
     }
 
-    compilerOptions.js_module_root.push(SRC_TEMP_DIR);
-
     const compilerOptionsArray = [];
     Object.keys(compilerOptions).forEach(function(option) {
       const value = compilerOptions[option];
@@ -376,11 +364,10 @@ function compile(
       }
     });
 
-    const gulpSrcs = convertPathsToTmpRoot(srcs);
-
     if (options.typeCheckOnly) {
       return gulp
-        .src(gulpSrcs, {base: SRC_TEMP_DIR})
+        .src(srcs, {base: '.'})
+        .pipe(preClosureBabel())
         .pipe(sourcemaps.init({loadMaps: true}))
         .pipe(gulpClosureCompile(compilerOptionsArray, checkTypesNailgunPort))
         .on('error', err => {
@@ -392,7 +379,8 @@ function compile(
     } else {
       timeInfo.startTime = Date.now();
       return gulp
-        .src(gulpSrcs, {base: SRC_TEMP_DIR})
+        .src(srcs, {base: '.'})
+        .pipe(preClosureBabel())
         .pipe(sourcemaps.init({loadMaps: true}))
         .pipe(gulpClosureCompile(compilerOptionsArray, distNailgunPort))
         .on('error', err => {