--- a/build-system/tasks/helpers.js
+++ b/build-system/tasks/helpers.js
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-const babel = require('@babel/core');
 const babelify = require('babelify');
 const browserify = require('browserify');
 const buffer = require('vinyl-buffer');
@@ -23,13 +22,11 @@ const conf = require('../compile/build.conf');
 const del = require('del');
 const file = require('gulp-file');
 const fs = require('fs-extra');
-const globby = require('globby');
 const gulp = require('gulp');
 const gulpIf = require('gulp-if');
 const gulpWatch = require('gulp-watch');
 const istanbul = require('gulp-istanbul');
 const log = require('fancy-log');
-const micromatch = require('micromatch');
 const path = require('path');
 const regexpSourcemaps = require('gulp-regexp-sourcemaps');
 const rename = require('gulp-rename');
@@ -37,11 +34,6 @@ const source = require('vinyl-source-stream');
 const sourcemaps = require('gulp-sourcemaps');
 const watchify = require('watchify');
 const wrappers = require('../compile/compile-wrappers');
-const {
-  BABEL_SRC_GLOBS,
-  SRC_TEMP_DIR,
-  THIRD_PARTY_TRANSFORM_GLOBS,
-} = require('../compile/sources');
 const {
   VERSION: internalRuntimeVersion,
 } = require('../compile/internal-version');
@@ -698,46 +690,6 @@ function toPromise(readable) {
   });
 }
 
-/**
- * @param {{isEsmBuild: boolean|undefined, isCheckTypes: boolean|undefined, isFortesting: boolean|undefined, isSinglePass: boolean|undefined}=} options
- */
-function transferSrcsToTempDir(options = {}) {
-  log(
-    'Performing pre-closure',
-    colors.cyan('babel'),
-    'transforms in',
-    colors.cyan(SRC_TEMP_DIR)
-  );
-  const files = globby.sync(BABEL_SRC_GLOBS);
-  const babelPlugins = conf.plugins({
-    isEsmBuild: options.isEsmBuild,
-    isSinglePass: options.isSinglePass,
-    isForTesting: options.isForTesting,
-    isChecktypes: options.isChecktypes,
-    isPostCompile: false,
-  });
-
-  files.forEach(file => {
-    if (
-      (file.startsWith('node_modules/') || file.startsWith('third_party/')) &&
-      !micromatch.isMatch(file, THIRD_PARTY_TRANSFORM_GLOBS)
-    ) {
-      fs.copySync(file, `${SRC_TEMP_DIR}/${file}`);
-      return;
-    }
-
-    const {code} = babel.transformFileSync(file, {
-      plugins: babelPlugins,
-      retainLines: true,
-      compact: false,
-    });
-    const name = `${SRC_TEMP_DIR}/${file}`;
-    fs.outputFileSync(name, code);
-    process.stdout.write('.');
-  });
-  console.log('\n');
-}
-
 module.exports = {
   applyAmpConfig,
   BABELIFY_GLOBAL_TRANSFORM,
@@ -757,5 +709,4 @@ module.exports = {
   printConfigHelp,
   printNobuildHelp,
   toPromise,
-  transferSrcsToTempDir,
 };