**Background:**

- AMP's minified build (`gulp dist`) and type checker (`gulp check-types`) run a set of `babel` transforms on source files before invoking closure compiler. 
- Today, transforms are run on **_all_** source files at once, after which the results are written to disk in a temp directory. Following this, closure compiler is invoked on the temp directory.
- This is particularly inefficient for flags like `--core_runtime_only` and `--noextensions`, which only compile a small subset of source files.

<img width="976" alt="Screen Shot 2020-03-26 at 12 28 24 AM" src="https://user-images.githubusercontent.com/26553114/77610359-e0c25b80-6ef8-11ea-95ac-367b392be724.png">

**This PR does the following:**
- Refactors pre-closure babel transforms to run directly on the `gulp.src` stream
- Moves all transform code to `build-system/compile/pre-closure-babel.js`
- Implements a caching mechanism so every source file is transformed at most once
- Deletes `transferSrcsToTempDir()` and all associated code
- Ensures that only those source files that are compiled are first transformed

**Performance improvements (Macbook Pro):**

Command | Before | After
--- | --- | ---
`gulp dist --core_runtime_only` | ~50s | ~30s
`gulp check-types` | ~1m 40s | ~1m 10s
`gulp dist --extensions_from examples/article.amp.html` | ~2m | ~1m 40s
`gulp dist` | ~7m | ~6m 30s
`gulp dist --single_pass` | ~5m | ~5m

**Coming up:**

- [ ] Enable a minified version of `gulp` with instant startup and lazy building


Addresses https://github.com/ampproject/amphtml/pull/26779#discussion_r397248222