@rsimha 2020-03-26
@erwinmombay @kristoferbaxter Extensively tested this locally, and it's now ready for a review. This will pave the way for minified lazy builds, and the pattern could potentially be adopted for integration tests as well (they currently do an all-at-once transform).